feat(firestore): rework firestore queries
